ftdi: Lattice JTAG HUB support
Use Lattice FPGA's JTAG hub encoding when connected to a CPU.
This is somewhat hackish, as it probably should not be implemented in
an interface driver, as the HUB is really not interface specific. But it
was the most forward way to achieve this for now.

The actual encoding mechanism was extracted from the lattice cableserver
openocd plugin.

Change-Id: I3167ef0b5dfeaa4b03f358d7b53eb63ae0c48f6e
Signed-off-by: Julian Scheel <julian@jusst.de>
Signed-off-by: Oskar Munz <omunz@jusst.de>